backend/requirements.txt
Flask
Flask-Migrate
Flask-Session
Flask-SQLAlchemy
python-dotenv
bcrypt
pg8000
email_validator